requirements.txt
amqp==5.2.0
bcrypt==3.2.0
billiard==4.2.0
blinker==1.8.2
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Deprecated==1.2.14
Flask==2.0.1
Flask-Cors==3.0.10
Flask-HTTPAuth==4.8.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.7.0
Flask-Mail==0.10.0
Flask-MySQL==1.5.2
gunicorn==22.0.0
httpagentparser==1.9.5
idna==3.7
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.3.7
limits==3.13.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.1
prompt_toolkit==3.0.47
pycparser==2.22
# pycrypto==2.6.1
pycryptodome==3.20.0
Pygments==2.18.0
PyJWT==2.8.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==0.19.1
redis2==2.10.6.3
requests==2.32.3
rich==13.7.1
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==2.2.2
wrapt==1.16.0
